+++ Encryption.php
@@ -25,8 +25,22 @@ final class Encryption extends AbstractEncryption
      * {@inheritdoc}
      *
      * @throw \Defuse\Crypto\Exception\EnvironmentIsBrokenException
+     *
+     * @deprecated since version 1.2, to be removed in 2.0. Use encryptWithKey instead
      */
     public function encrypt($plaintext, KeyInterface $key)
+    {
+        @trigger_error('encrypt() is deprecated since version 1.2 and will be removed in 2.0. Use encryptWithKey instead.', E_USER_DEPRECATED);
+
+        return BaseCrypto::encrypt($plaintext, $this->unlockKey($key));
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @throw \Defuse\Crypto\Exception\EnvironmentIsBrokenException
+     */
+    public function encryptWithKey($plaintext, KeyInterface $key)
     {
         return BaseCrypto::encrypt($plaintext, $this->unlockKey($key));
     }
@@ -36,8 +50,23 @@ final class Encryption extends AbstractEncryption
      *
      * @throws \Defuse\Crypto\Exception\EnvironmentIsBrokenException
      * @throws \Defuse\Crypto\Exception\WrongKeyOrModifiedCiphertextException
+     *
+     * @deprecated since version 1.2, to be removed in 2.0. Use decryptWithKey instead
      */
     public function decrypt($ciphertext, KeyInterface $key)
+    {
+        @trigger_error('decrypt() is deprecated since version 1.2 and will be removed in 2.0. Use decryptWithKey instead.', E_USER_DEPRECATED);
+
+        return BaseCrypto::decrypt($ciphertext, $this->unlockKey($key));
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @throws \Defuse\Crypto\Exception\EnvironmentIsBrokenException
+     * @throws \Defuse\Crypto\Exception\WrongKeyOrModifiedCiphertextException
+     */
+    public function decryptWithKey($ciphertext, KeyInterface $key)
     {
         return BaseCrypto::decrypt($ciphertext, $this->unlockKey($key));
     }
@@ -47,8 +76,23 @@ final class Encryption extends AbstractEncryption
      *
      * @throws \Defuse\Crypto\Exception\IOException
      * @throws \Defuse\Crypto\Exception\EnvironmentIsBrokenException
+     *
+     * @deprecated since version 1.2, to be removed in 2.0. Use encryptFileWithKey instead
      */
     public function encryptFile($inputFilename, $outputFilename, KeyInterface $key)
+    {
+        @trigger_error('encryptFile() is deprecated since version 1.2 and will be removed in 2.0. Use encryptFileWithKey instead.', E_USER_DEPRECATED);
+
+        BaseCryptoFile::encryptFile($inputFilename, $outputFilename, $this->unlockKey($key));
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @throws \Defuse\Crypto\Exception\IOException
+     * @throws \Defuse\Crypto\Exception\EnvironmentIsBrokenException
+     */
+    public function encryptFileWithKey($inputFilename, $outputFilename, KeyInterface $key)
     {
         BaseCryptoFile::encryptFile($inputFilename, $outputFilename, $this->unlockKey($key));
     }
@@ -59,8 +103,24 @@ final class Encryption extends AbstractEncryption
      * @throws \Defuse\Crypto\Exception\IOException
      * @throws \Defuse\Crypto\Exception\EnvironmentIsBrokenException
      * @throws \Defuse\Crypto\Exception\WrongKeyOrModifiedCiphertextException
+     *
+     * @deprecated since version 1.2, to be removed in 2.0. Use decryptFileWithKey instead
      */
     public function decryptFile($inputFilename, $outputFilename, KeyInterface $key)
+    {
+        @trigger_error('decryptFile() is deprecated since version 1.2 and will be removed in 2.0. Use decryptFileWithKey instead.', E_USER_DEPRECATED);
+
+        BaseCryptoFile::decryptFile($inputFilename, $outputFilename, $this->unlockKey($key));
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @throws \Defuse\Crypto\Exception\IOException
+     * @throws \Defuse\Crypto\Exception\EnvironmentIsBrokenException
+     * @throws \Defuse\Crypto\Exception\WrongKeyOrModifiedCiphertextException
+     */
+    public function decryptFileWithKey($inputFilename, $outputFilename, KeyInterface $key)
     {
         BaseCryptoFile::decryptFile($inputFilename, $outputFilename, $this->unlockKey($key));
     }